atc: always emit a metric for each worker state
This commit also backfills tests for ths WorkersState metric. It turns out that
there was enough untested logic there that a bug was introduced when
concourse/concourse#4895 was implemented, so it's high time to have some tests!

It was more difficult than I expected to get the metric into a test harness --
the code had a problem with what Michael Feathers calls "sensing". It's not
expensive to instantiate the metric, but it is difficult to observe the effect
of its `Emit` method being called, since the `emitLoop` function in the `metric`
package asynchronously passes events to the emitter in a goroutine.

I also think its possible that running these tests in multiple threads could
case data races -- accordingly I tried to put pretty aggressive names on the
helper functions. I think we should be able to isolate the metric from the
`emissions` channel in tests, but I think it would be best to do the same thing
uniformly across all the various metrics. I haven't really thought it through
yet at this point.

concourse/concourse#4985
concourse/concourse#3856

Signed-off-by: Jamie Klassen <cklassen@pivotal.io>